backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
python-multipart>=0.0.9,<1
pydantic>=2.6,<3
pdfplumber>=0.11,<1
openai>=1.40,<2
python-dotenv>=1.0,<2
